backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
reportlab==3.6.8
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Flask-Cors==3.0.10
gunicorn==20.1.0
unidecode==1.3.6
ldap3==2.9.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pysnmp==4.4.12
pyasn1==0.4.8
pycryptodome==3.21.0
PyJWT==2.8.0
    